tercer commit

+++ app.py
@@ -1,22 +1,22 @@
 from flask import Flask, render_template,request
 
-def F_P(tasa: float, n: int ):
+def F_P(tasa: float, n: float ):
     return (1+tasa)**n
     
-def P_F(tasa: float, n: int):
+def P_F(tasa: float, n: float):
     return (1+tasa)**(-n)
 
-def F_A(tasa: float, n: int):
+def F_A(tasa: float, n: float):
     temp = ((1+tasa)**n - 1)/(tasa)
     return temp
-def A_F(tasa: float, n: int):
+def A_F(tasa: float, n: float):
     return F_A(tasa,n)**(-1)
 
-def P_G( tasa: float, n: int):
+def P_G( tasa: float, n: float):
     temp = ((1+tasa)**n - tasa*n-1)/(tasa**2* (1+tasa)**n)
     return temp
 
-def A_G(tasa: float, n: int):
+def A_G(tasa: float, n: float):
     temp = tasa**(-1)  - n/(  (1+tasa)**n -1    )
     return temp
